ai_assistant/requirements.txt
litellm




pymupdf
pdfplumber
python-docx
openpyxl
opencv-python
pytesseract


